id and userId changed to int64

+++ internal/post-crud-service/domain/post_domain.go
@@ -1,8 +1,8 @@
 package domain
 
 type Post struct {
-	Id     int    `json:"id"`
-	UserId int    `json:"userId"`
+	Id     int64  `json:"id"`
+	UserId int64  `json:"userId"`
 	Title  string `json:"title"`
 	Body   string `json:"body"`
 }
